Load all fixtures in first Test 'LoginFormTest' Remove load fixtures from 'MoqupTest'
